requirements.txt
httpx>=0.24.0
azure-identity>=1.12.0
mcp[cli]>=0.0.17